+++ lib/grinder.dart
@@ -1,3 +1,6 @@
 library bwu_utils.grinder;
 
-export 'grinder/grinder_content_shell_drt.dart';
+export 'grinder/content_shell_drt.dart';
+export 'grinder/linter_task.dart';
+export 'grinder/analyzer_task.dart';
+export 'grinder/dartformat_task.dart';