+++ lib/testing_server.dart
@@ -1,3 +1,5 @@
 library bwu_utils.testing_server;
 
 export 'testing/server.dart';
+export 'bwu_utils_server.dart';
+