+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## 0.1.0
 
 - Merged bwu_utils_server, bwu_utils_browser
-- Added Grunt tasks
+- Added Grunt tasks (analyzer, dartformat, linter)
+- Added Grund content_shell browser implementation
 - Added some utils for unit testing
+- renamed parseIntFromStringWithUnit in browser.parse_int to parseIntDropUnit
+- use unittest 0.12.0-alpha.1